[#16894] - refactoring to use fakes
