+++ tests/unit/Html/Helper/Input/Select/ResultsetDataTest.php
@@ -13,15 +13,10 @@ declare(strict_types=1);
 
 namespace Phalcon\Tests\Unit\Html\Helper\Input\Select;
 
-use ArrayIterator;
-use BadMethodCallException;
-use Closure;
 use InvalidArgumentException;
 use Phalcon\Html\Helper\Input\Select\ResultsetData;
-use Phalcon\Messages\MessageInterface;
-use Phalcon\Mvc\Model\ResultsetInterface;
-use Phalcon\Mvc\ModelInterface;
 use Phalcon\Tests\AbstractUnitTestCase;
+use Phalcon\Tests\Unit\Html\Helper\Input\Select\Fake\FakeResultset;
 
 final class ResultsetDataTest extends AbstractUnitTestCase
 {
@@ -34,10 +29,7 @@ final class ResultsetDataTest extends AbstractUnitTestCase
         $this->expectException(InvalidArgumentException::class);
         $this->expectExceptionMessage('requires exactly two values');
 
-        new ResultsetData(
-            $this->buildResultsetMock([]),
-            ['id']
-        );
+        new ResultsetData(new FakeResultset([]), ['id']);
     }
 
     /**
@@ -46,8 +38,7 @@ final class ResultsetDataTest extends AbstractUnitTestCase
      */
     public function testGetOptionsReturnsEmptyArrayForEmptyResultset(): void
     {
-        $resultset = $this->buildResultsetMock([]);
-        $data      = new ResultsetData($resultset, ['id', 'name']);
+        $data = new ResultsetData(new FakeResultset([]), ['id', 'name']);
 
         $this->assertSame([], $data->getOptions());
     }
@@ -63,10 +54,9 @@ final class ResultsetDataTest extends AbstractUnitTestCase
             ['id' => '2', 'name' => 'Ford'],
         ];
 
-        $resultset = $this->buildResultsetMock($rows);
-        $data      = new ResultsetData($resultset, ['id', 'name']);
-
+        $data     = new ResultsetData(new FakeResultset($rows), ['id', 'name']);
         $expected = ['1' => 'Ferrari', '2' => 'Ford'];
+
         $this->assertSame($expected, $data->getOptions());
     }
 
@@ -88,93 +78,9 @@ final class ResultsetDataTest extends AbstractUnitTestCase
         $row2->method('readAttribute')
              ->willReturnMap([['id', '2'], ['name', 'Ford']]);
 
-        $resultset = $this->buildResultsetMock([$row1, $row2]);
-
-        $data   = new ResultsetData($resultset, ['id', 'name']);
+        $data   = new ResultsetData(new FakeResultset([$row1, $row2]), ['id', 'name']);
         $actual = $data->getOptions();
 
         $this->assertSame(['1' => 'Ferrari', '2' => 'Ford'], $actual);
-    }
-
-    /**
-     * @param array $rows
-     *
-     * @return ResultsetInterface
-     */
-    private function buildResultsetMock(array $rows): ResultsetInterface
-    {
-        return new class ($rows) extends ArrayIterator implements ResultsetInterface {
-            public function __construct(array $rows)
-            {
-                parent::__construct($rows);
-            }
-
-            public function delete(Closure | null $conditionCallback = null): bool
-            {
-                throw new BadMethodCallException('Not implemented');
-            }
-
-            public function filter(callable $filter): array
-            {
-                throw new BadMethodCallException('Not implemented');
-            }
-
-            public function getCache(): mixed
-            {
-                throw new BadMethodCallException('Not implemented');
-            }
-
-            public function getFirst(): mixed
-            {
-                throw new BadMethodCallException('Not implemented');
-            }
-
-            public function getHydrateMode(): int
-            {
-                throw new BadMethodCallException('Not implemented');
-            }
-
-            public function getLast(): ModelInterface | null
-            {
-                throw new BadMethodCallException('Not implemented');
-            }
-
-            public function getMessages(): array
-            {
-                throw new BadMethodCallException('Not implemented');
-            }
-
-            public function getType(): int
-            {
-                throw new BadMethodCallException('Not implemented');
-            }
-
-            public function isFresh(): bool
-            {
-                throw new BadMethodCallException('Not implemented');
-            }
-
-            public function setHydrateMode(int $hydrateMode): ResultsetInterface
-            {
-                throw new BadMethodCallException('Not implemented');
-            }
-
-            public function setIsFresh(bool $isFresh): ResultsetInterface
-            {
-                throw new BadMethodCallException('Not implemented');
-            }
-
-            public function toArray(): array
-            {
-                throw new BadMethodCallException('Not implemented');
-            }
-
-            public function update(
-                mixed $data,
-                Closure | null $conditionCallback = null
-            ): bool {
-                throw new BadMethodCallException('Not implemented');
-            }
-        };
     }
 }
